Lab 2
For this lab, you will upload two files: Your RScript file and a write-up (in Word or LaTeX) of all your answers to each problem. If a problem asks for a visualization (e.g. a graph) or a table, be sure to include an image of it in your write-up. Save your RScript file as your first and last name:  firstname.lastname.R. Your write-up should be saved in a similar fashion.